plugin: move endpoints defs to separate module
